requirements.txt
atomicwrites==1.4.1
attrs==23.2.0
blinker==1.7.0
build==1.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cryptography==42.0.5
Flask==3.0.2
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.17.3
jsonschema-spec==0.1.6
kafka-python==2.0.2
lazy-object-proxy==1.10.0
lz4==4.3.3
MarkupSafe==2.1.5
more-itertools==10.2.0
openapi-core==0.16.0
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.4
packaging==23.2
parse==1.20.1
pathable==0.4.3
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tapipy==1.6.3
typing_extensions==4.10.0
urllib3==1.26.18
Werkzeug==3.0.1
